refactor: Move Package Widget Items
