+++ feature/main/src/main/res/values/dimens.xml
@@ -31,6 +31,12 @@
     <dimen name="prep_dialog_dropdown_button_height">48dp</dimen>
     <dimen name="prep_dialog_dropdown_item_height">200dp</dimen>
 
+    <dimen name="recipe_card_item_width">150dp</dimen>
+    <dimen name="recipe_card_item_height">188dp</dimen>
+    <dimen name="recipe_card_item_blur">12dp</dimen>
+    <dimen name="recipe_card_item_image_height">160dp</dimen>
+    <dimen name="recipe_card_item_box_height">28dp</dimen>
+
     <dimen name="drawer_page_width">300dp</dimen>
     <dimen name="drawer_page_padding">20dp</dimen>
     <dimen name="drawer_vertical_padding">8dp</dimen>
